# EnumExtensions.GetDescription Micro-Optimization