--- a/JikanDotNet/Extensions/EnumExtensions.cs
+++ b/JikanDotNet/Extensions/EnumExtensions.cs
@@ -1,16 +1,13 @@
 using System;
 using System.ComponentModel;
-using System.Linq;
 using System.Reflection;
 
 namespace JikanDotNet.Extensions
 {
 	internal static class EnumExtensions
 	{
-		public static string GetDescription(this Enum source) => source
-					.GetType()
-					.GetMember(source.ToString())
-					.FirstOrDefault()
+		public static string GetDescription<T>(this T source) where T : Enum => typeof(T)
+					.GetField(source.ToString())
 					?.GetCustomAttribute<DescriptionAttribute>()
 					?.Description;
 	}