From my benchmarks, this appears to be ~25% faster and allocate a bit less memory per call

|                Method |        Mean |     Error |    StdDev | Allocated |
|---------------------- |------------:|----------:|----------:|----------:|
|   GetDescriptionJikan | 1,701.56 ns | 32.576 ns | 54.427 ns |     288 B |
| GetDescriptionGeneric | 1,283.37 ns | 25.371 ns | 24.918 ns |     256 B |